WIP: DEBUG: create a test harness for mmiocpy
the arm memcpy_fromio and memcpy_toio were changed in 2015 to use
mmiocpy which is a function alias (same bytes defined for the function)
to memcpy, but avoids gcc optimizations.

Create a wrapper to test it works on defined memory.  The buffer
is actually 1 order above the const order symbol; the first half
is the read buffer and the second is the source buffer.

Both source and destination start out as chr(1..255) repeated
and you can do just copys, copys after a fill, or copys after a
base refresh of the base buffer (a zero length write).

Set the parameters (set = do memset or cpy), then echo to cpy
to trigger the copy (first char of cpy is the value to fill,
or zero length write does copy).