WIP: aspeed-smc: use io string accessors for fifo
The memcpy_{to,from}io implemented via mmiocpy on arm appear to
stutter when switching from words back to bytes at the end of
the copy. While we debate if this is acceptable for them and/or
design the patch fixup (near the b 13b in the copy_forward macro
in copy_template.S) lets make a fifo accessor that will be nice
to our fifo.

This has some horrible hacks to make inw/outw work too, like
selecting a PCMCIA intermeditate symbol to get IO_SPACE_LIMIT
since multi-platform doesn't have an mach/io.h to set it.

NOTE: while we do have an mach/aspeed dir I have a patch to
move that one file back to the directory with its only
include site like all the other multiplatform port so we
can go upstream.  Or we remove it altogether with a dts based
pinmux :-).

Testing-Status: Crashes by blowing return address back to spi-nor-scan
Signed-off-by: Milton Miller <miltonm@us.ibm.com>