WIP: fix aspeed_smc_from_fifo
A copy and paste error meant that we did the initial alignment
bytes a second time after the main transfer instead of the remaining
byte length.   The result was seen during boot as a stack smash and
segv on return address from spi_nor_scan().

This now boots with both fifo_from and fifo_to true.