pinctrl: Add pinctrl-aspeed, core pinctrl support for Aspeed SoC
Add pinctrl/pinmux support for the Aspeed SoCs. Configuration of the
Aspeed SoCs is somewhat intricate: Enabling functions can involve writes
of multiple bits to multiple registers, and signals provided by a pin
are determined on a priority basis. That is, without care, it's possible
to configure a function in the mux and to not gain expected
functionality.

Note that this doesn't implement the pinctrl/pinmux/pinconf for any
particular SoC, just adds the framework to enable defining the features.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>