slab: Force SLAB_RED_ZONE for cache line aligned sub-page objects
When FORCE_DEBUG is in effect (via CONFIG_SLAB_DEBUG) and the requested
alignment is 1 cache line, reserve two cache lines from each slab and
force SLAB_RED_ZONE and SLAB_STORE_USER for objects that will continue
to fit under the 4k + alignment barrier.

The default alignment for the kmalloc caches is SLAB_HWCACHE_ALIGNED.
Architectures with non-coherent DMA require this to allow dma to objects
allocated via kmalloc.  However, CONFIG_SLAB_DEBUG disables SLAB_RED_ZONE
and SLAB_STORE_USER if the object alignment is greater than that required
for an unsigned long long.  This means all the kmalloc caches and many
other objects will not have red zoning enabled.

The above limitation was hindering the debug of a slab overrun of the
kmalloc-256 slab.  CONFIG_SLAB_DEBUG showed the first byte of an object
was being overritten but the previous object was either stored or text
with trailing Z indicating the object had been freed and reused.  By
adding 1 cache line to each object size, the red zone quickly identified
the function freeing the object and the process name in the WARN provided
the additional details needed to track down the slab overrun.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>