mtd: aspeed-smc: remove ioport abuse
When the memcpy_{to,from}_io were found to stutter on little
endian arm the io string functions were chosen based on the low
level assembly being closest to the optimal choice.  However,
it was discovered the generic usage required ugly workarounds to
avoid the arm io port resource limits that try to avoid bad io
ports from being used and the mmio resources could not be used
as iomap tokens.

Replace with read and write mmio routines and handle the memory
buffer address in C code.  Update the comment for why the
routines exist.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>