devicetree: update binding reg and add hardware description
The big change is the chip select segemnts are now managed by
the device driver and the device tree binding always shows one
single resource, regardless of the chips attached.

This means the chip dtsi files will get the reg property and not
need to be overidden with the details of the attached device,
and removes a possible mismatch between the device tree and the
programmed state of the hardware (either through hardare init
defaults or the bootloader).

This marks a few of the properties from common SPI bindings as
"not yet checked".  Actually the ast2500 dropped support for
mode 3 so clock phase and polarity are not changable on 2500,
and on 2400 they are linked.

This might not be the right place to give the background of the
hardware, but I wrote it here first.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>