spi-nor/aspeed-smc: Manage the segment address registers in the driver
This changes the aspeed mtd driver to manage the segment address
regisers from the driver, this eliminating a source of mismatch
and placing the kernel in control using its knowledge instead of
relying on hardware defaults or bootloaders to choose good values
for the windows based on the attached chips.

The initial commit factors out discovering the current window and
sanitizing the registers to avoid overlap.  The intent is to probe
the attached flash to discover their size from their jedec id and
adjust the windows to the required size.  Also planned is to enable
direct memory mapped access and dma access, at least for read
accesses when the flash size is supported by the hardware.

The spi-nor layer already trusts its database of jedec ids over any
information in the device tree, this continues that philosophy
by using the knowledge to optimize the programming of the hardware.
The device tree may be used in the future to limit the choices,
for example specifing a max clock rate based on board wiring.

This does not have any locking around sanitizing the segments
or copying the new addresses back to the per-chip base address.
In the present code registers are updated before the devices
are probed.

We may want to cache the offsets in the per-chip structure to
when we support DMA.

Signed-off-by: Milton Miller