WIP: convert palmetto dts to new spi-nor binding
This changes the reg field to have the full window available.

This should change the dtsi and move the fields around.

The fmc and smc names should be generic flash-controller labels
and palaced in the device tree.

The flash-layout should be adapted to the udpated partition binding that
places all partitions under a single subnode.

Applying the change will require CONFIG_VMSPLIT_2G to reserve enogh
vmap space for the iomap.

Caution: when I changed that config PHYS_OFFSET was changed but
the object/include/config/page/offset.h file was not touched and
many files were not rebuilt.  You may need to touch that or make
clean as a workaround.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>