[RFC] watchdog: print bootstatus in hex like status
Change the bootstatus to print 0x%x instead of %u.
This makes it match status, which was already in hex.

The status is a collection of bits.  The value for "Card has previously
reset the system" GPIOF_CARDREET is 32, this prints it as 0x20.  If the
card is not configured to reset, the status is bitwise or'd with 0x400
(GPIOF_ALARM).  An alarm that has been pinged and triggered will print
0x8420 in status then return 0x420.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>
---
A search for userspace impact has not been performed.